a round of cleanups courtesy of PVS studio

+++ Code/Geometry/UniformGrid3D.cpp
@@ -156,8 +156,9 @@ Point3D UniformGrid3D::getGridPointLoc(unsigned int pointId) const {
   Point3D res;
   res.x = (pointId % d_numX) * d_spacing;
   // the rounding here is intentional, we want the coordinates of a grid point
-  res.y = ((pointId % (d_numX * d_numY)) / d_numX) * d_spacing;
-  res.z = (pointId / (d_numX * d_numY)) * d_spacing;
+  res.y =
+      static_cast<double>((pointId % (d_numX * d_numY)) / d_numX) * d_spacing;
+  res.z = static_cast<double>(pointId / (d_numX * d_numY)) * d_spacing;
   res += d_offSet;  // d_origin;
   return res;
 }
